# Conflicts:
#	cocos/2d/CCScene.cpp
#	tests/cpp-tests/Classes/Scene3DTest/Scene3DTest.cpp

+++ cocos/base/CCDirector.cpp
@@ -1354,7 +1354,7 @@ void DisplayLinkDirector::stopAnimation()
     _invalid = true;
 }
 
-void DisplayLinkDirector::setAnimationInterval(double interval)
+void DisplayLinkDirector::setAnimationInterval(float interval)
 {
     _animationInterval = interval;
     if (! _invalid)
